// frontend/src/api/model/clinicalNoteKind.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type ClinicalNoteKind = typeof ClinicalNoteKind[keyof typeof ClinicalNoteKind];


export const ClinicalNoteKind = {
  EVOLUTION: 'EVOLUTION',
  OBSERVATION: 'OBSERVATION',
  INTERCONSULTATION: 'INTERCONSULTATION',
  NURSING: 'NURSING',
} as const;